使用 flyway 数据库自动化迁移工具
- 使用 flyway 进行数据库自动初始化
- 重构爬虫逻辑，优化对链接加入已处理，和从链接池中删除的逻辑
- 提取 DAO 接口，为其他数据库的接入作准备
